# Scaffolding to generate call_validation.rb via Bazel