--- a/gems/sorbet-runtime/BUILD
+++ b/gems/sorbet-runtime/BUILD
@@ -47,8 +47,8 @@ cc_binary(
     }),
     visibility = ["//tools:__pkg__"],
     deps = [
-        "@cxxopts",
         "//common",
+        "@cxxopts",
     ],
 )