--- a/gems/sorbet-runtime/lib/types/private/methods/_methods.rb
+++ b/gems/sorbet-runtime/lib/types/private/methods/_methods.rb
@@ -425,32 +425,17 @@ module T::Private::Methods
       @old_hooks.each(&:restore)
       @old_hooks = nil
     else
-      if T::Configuration::AT_LEAST_RUBY_2_7
-        old_included = T::Private::ClassUtils.replace_method(Module, :included) do |arg|
-          old_included.bind_call(self, arg)
-          ::T::Private::Methods._hook_impl(arg, arg.ancestors, self)
-        end
-        old_extended = T::Private::ClassUtils.replace_method(Module, :extended) do |arg|
-          old_extended.bind_call(self, arg)
-          ::T::Private::Methods._hook_impl(arg, arg.singleton_class.ancestors, self)
-        end
-        old_inherited = T::Private::ClassUtils.replace_method(Class, :inherited) do |arg|
-          old_inherited.bind_call(self, arg)
-          ::T::Private::Methods._hook_impl(arg, arg.ancestors, self)
-        end
-      else
-        old_included = T::Private::ClassUtils.replace_method(Module, :included) do |arg|
-          old_included.bind(self).call(arg)
-          ::T::Private::Methods._hook_impl(arg, arg.ancestors, self)
-        end
-        old_extended = T::Private::ClassUtils.replace_method(Module, :extended) do |arg|
-          old_extended.bind(self).call(arg)
-          ::T::Private::Methods._hook_impl(arg, arg.singleton_class.ancestors, self)
-        end
-        old_inherited = T::Private::ClassUtils.replace_method(Class, :inherited) do |arg|
-          old_inherited.bind(self).call(arg)
-          ::T::Private::Methods._hook_impl(arg, arg.ancestors, self)
-        end
+      old_included = T::Private::ClassUtils.replace_method(Module, :included) do |arg|
+        old_included.bind(self).call(arg)
+        ::T::Private::Methods._hook_impl(arg, arg.ancestors, self)
+      end
+      old_extended = T::Private::ClassUtils.replace_method(Module, :extended) do |arg|
+        old_extended.bind(self).call(arg)
+        ::T::Private::Methods._hook_impl(arg, arg.singleton_class.ancestors, self)
+      end
+      old_inherited = T::Private::ClassUtils.replace_method(Class, :inherited) do |arg|
+        old_inherited.bind(self).call(arg)
+        ::T::Private::Methods._hook_impl(arg, arg.ancestors, self)
       end
       @old_hooks = [old_included, old_extended, old_inherited]
     end