--- a/gems/sorbet-runtime/lib/types/private/methods/call_validation_2_6.rb
+++ b/gems/sorbet-runtime/lib/types/private/methods/call_validation_2_6.rb
@@ -1,1 +1,1 @@
-# Hello, world!
+# Hello, world! --bind-call=false