--- a/gems/sorbet-runtime/lib/types/private/methods/call_validation_2_7.rb
+++ b/gems/sorbet-runtime/lib/types/private/methods/call_validation_2_7.rb
@@ -1,1 +1,1 @@
-# Hello, world!
+# Hello, world! --bind-call=true