--- a/tools/scripts/update_exp_files.sh
+++ b/tools/scripts/update_exp_files.sh
@@ -16,4 +16,5 @@ gems/sorbet/test/hidden-method-finder/update_hidden_methods_exp.sh "$@"
 
 ./bazel test \
   //gems/sorbet/test/snapshot:update \
+  //gems/sorbet-runtime:update_call_validation \
   //test/cli:update //test/lsp:update -c opt "$@"