<!-- (optional) Explain your change, focusing on the details of the solution. This is a great place to call out user-visible changes. -->


### Motivation
<!-- Why make this change? Describe the problem, not the solution. This can also be a link to an issue. -->

We have a bunch of pseudo generated code in call_validation.rb (all the fast
paths).

We also want to make a change to use `UnboundMethod#bind_call` if available,
with zero overhead (i.e., if you're using Ruby 2.7 or greater, it's as if the
whole project used it exclusively, and if you're on 2.6 or before,
sorbet-runtime still works, with no required monkey patches).

@djudd-stripe has done some work to measure the overhead of various `bind_call`
alternatives:

```
master:                                               476.321 ns
2.6/2.7 compatibility wrappers:                       384.027 ns
use bind_call exclusively, breaking compatibility:    343.558 ns
```

The plan to get to the best-case scenario below, while maintaining
compatibility is to generate two `call_validation.rb` files, and require one or
the other by reflecting on the current Ruby version.

**Thus**: this PR sets up the Bazel scaffolding to generate
`call_validation.rb`.

### Next steps:

tbd whether these should be in this change, a stacked change, or an unstacked future change.

- [ ] Get to parity (make `generate_call_validation.cc` emit the same thing that's in call_validation.rb now)
- [ ] Update the generator to take advantage of `bind_call`.
- [ ] (optional) refactor it so that the various fast paths are generated too


### Test plan
<!-- If you did not write tests for this change, replace the message below explaining why not. Why we should be confident this change is correct? If you changed the website, please include a screenshot of the proposed changes. -->

I committed a failing test to make sure that Buildkite fails:

<https://buildkite.com/sorbet/sorbet/builds/14915#ed3756ba-1838-4cb7-a88f-1c4d99500c39/254-257>

I also verified that the new `update` commands work locally, as written in the
README.