@jez 2021-02-23
> Do we think that `bind_call` taking up slightly more stack space than `bind().call` (even given the fewer allocations) is likely to cause problems?

What do you mean slightly more stack space? I was not aware of anything taking more stack space.

The thing that took too much stack space last time was when we had compatibility wrappers that added another frame to the call stack, but I think that the approach of having two files will be mean that the stack space is unchanged.

> This is basically #3683 with less sed, more bazel, and more tightly scoped? I do like having everything in bazel.

Yes, and also I think that it will be easier to use this scaffolding to reduce all the code duplication that already exists in call_validation (e.g., we could generate the 10 fast-path validators).