@jez 2021-02-23
> with less sed

Also to be clear: I think it was great that we used `sed` at first just to quickly check the approach overall. Thanks for putting that together!