@froydnj 2021-02-23
> > Do we think that `bind_call` taking up slightly more stack space than `bind().call` (even given the fewer allocations) is likely to cause problems?
> 
> What do you mean slightly more stack space? I was not aware of anything taking more stack space.
> 
> The thing that took too much stack space last time was when we had compatibility wrappers that added another frame to the call stack, but I think that the approach of having two files will be mean that the stack space is unchanged.

Mmmm, compat wrappers, yes.  I was thinking actual C stack usage:

https://github.com/ruby/ruby/blob/931815bfd86df603337194f3fcefb46bfe3e7940/proc.c#L2472-L2488

since we're putting `struct METHOD` on the stack now...which might matter?  (I mean, we should try to make it not matter, but maybe it does right now?)

> > This is basically #3683 with less sed, more bazel, and more tightly scoped? I do like having everything in bazel.
> 
> Yes, and also I think that it will be easier to use this scaffolding to reduce all the code duplication that already exists in call_validation (e.g., we could generate the 10 fast-path validators).

:+1:

> > with less sed
> 
> Also to be clear: I think it was great that we used `sed` at first just to quickly check the approach overall. Thanks for putting that together!

:+1: