@jez 2021-02-28
After implementing the code generation, here's a benchmark run:

Tested on Ryzen 3700X, ruby 2.7.2p137

    master:                                               sig {params(x: Integer).void}: 422.979 ns
    2.6/2.7 compatibility wrappers:                       sig {params(x: Integer).void}: 367.138 ns
    use bind_call exclusively, breaking compatibility:    sig {params(x: Integer).void}: 322.575 ns
    this PR:                                              sig {params(x: Integer).void}: 304.047 ns

This PR is technically 20ns faster than just throwing away compatibility and using `bind_call` exclusively, so there's probably a reasonable amount of noise here.

But approximately 100ns reduction (39% faster), and as fast as breaking compatibility.